problem 12 needs rewriting with a MUCH quicker method

+++ optimise/problem012.cpp
@@ -11,24 +11,24 @@ void debugOut(string s)
     }
 }
 bool *isPrime;
-bool *getFactorsSieve(int number)
+bool *getFactorsSieve(long number)
 {
-    int max = (number) + 1;
-    int sq = sqrt(max) + 1;
+    long max = (number) + 1;
+    long sq = sqrt(max) + 1;
     bool *isFactor = new bool[max];
-    for (int i = 0; i != max; i++)
+    for (long i = 0; i != max; i++)
     {
         isFactor[i] = false;
     }
     isFactor[1] = isFactor[number] = true;
-    for (int i = 2; i != sq; i++)
+    for (long i = 2; i != sq; i++)
     {
         if (isPrime[i])
         {
             if (number % i == 0)
             {
                 isFactor[i] = true; 
-                for (int j = 2 * i; j < max; j+= i)
+                for (long j = 2 * i; j < max; j+= i)
                 {
                     if (number % j == 0)
                     {
@@ -45,12 +45,12 @@ int main()
     cout << "Finding primes" << endl;
     isPrime = getSieve(1000000000);
     cout << "Found primes" << endl;
-    int top = 0;
-    for(int number = 3, last = 3; ;number += last, last++)
+    long top = 0;
+    for(long number = 3, last = 3; ;number += last, last++)
     {
-        int count = 0;
+        long count = 0;
         bool *isFactor = getFactorsSieve(number);
-        for (int i = 0; i != (number) + 1; i++)
+        for (long i = 0; i != (number) + 1; i++)
         {
             if (isFactor[i])
             {
